Add optional binary path for HandbrakeCLI
Semi-pulled in changes from @BurakBoz's fork, this adds the option to
add a handbrakeBinary property on the Config class. But also doesn't
implement the checks for PHP_OS being Windows. I've tested this on
Windows and it didn't appear to be necessary because Windows should
detect .exe files when no extension is provided. As long as it's in the
path. The command string will default to 'HandbrakeCLI' if no path is
provided in the config.